refactor: 테마별 예약 가능 시간 조회 API 엔드포인트 변경 (/times 아래)
