--- src/main/java/roomescape/theme/controller/dto/ThemeAvailableTimeResponseDto.java
@@ -1,10 +1,0 @@
-package roomescape.theme.controller.dto;
-
-import roomescape.reservation.domain.ReservationTime;
-
-public record ThemeAvailableTimeResponseDto(Long id, String startAt) {
-    public static ThemeAvailableTimeResponseDto from(ReservationTime time) {
-        return new ThemeAvailableTimeResponseDto(time.getId(), time.getStartAt());
-    }
-}
-
